BE:fix: add error handling about validating meetup startAt time

+++ backend/service/src/main/java/org/lightnsalt/hikingdom/service/club/service/meetup/MeetupMemberServiceImpl.java
@@ -1,5 +1,6 @@
 package org.lightnsalt.hikingdom.service.club.service.meetup;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -45,6 +46,9 @@ public class MeetupMemberServiceImpl implements MeetupMemberService {
 		final Meetup meetup = meetupRepository.findById(meetupId)
 			.orElseThrow(() -> new GlobalException(ErrorCode.MEETUP_NOT_FOUND));
 
+		if (meetup.getStartAt().isBefore(LocalDateTime.now()))
+			throw new GlobalException(ErrorCode.MEETUP_ALREADY_DONE);
+
 		meetupMemberRepository.save(MeetupMember.builder().meetup(meetup).member(member).build());
 	}
 
@@ -54,7 +58,12 @@ public class MeetupMemberServiceImpl implements MeetupMemberService {
 		final Member member = memberRepository.findByEmail(email)
 			.orElseThrow(() -> new GlobalException(ErrorCode.MEMBER_UNAUTHORIZED));
 		final MeetupMember meetupMember = meetupMemberRepository.findByMeetupIdAndMemberId(meetupId,
-			member.getId()).orElseThrow(() -> new GlobalException(ErrorCode.MEMBER_UNAUTHORIZED));
+			member.getId()).orElseThrow(() -> new GlobalException(ErrorCode.MEETUP_NOT_JOINED));
+		final Meetup meetup = meetupRepository.findById(meetupId)
+			.orElseThrow(() -> new GlobalException(ErrorCode.MEETUP_NOT_FOUND));
+
+		if (meetup.getStartAt().isBefore(LocalDateTime.now()))
+			throw new GlobalException(ErrorCode.MEETUP_ALREADY_DONE);
 
 		meetupMemberRepository.delete(meetupMember);
 	}
@@ -62,7 +71,6 @@ public class MeetupMemberServiceImpl implements MeetupMemberService {
 	@Override
 	@Transactional
 	public List<MeetupMemberDetailListRes> findMeetupMemberDetail(Long clubId, Long meetupId) {
-
 		// 형 변환
 		return getMeetupMember(clubId, meetupId).stream()
 			.map(meetupMember -> new MeetupMemberDetailListRes(meetupMember.getMember()))
